+++ example/starter_project/lib/starter_project.dart
@@ -1,5 +1,3 @@
-import 'item.dart';
-
 /// Non refactored `GildedRose` class from Gilded Rose Kata
 final class GildedRose {
   final List<Item> items;
@@ -62,4 +60,16 @@ final class GildedRose {
       }
     }
   }
+}
+
+/// `Item` class for representing items in the Gilded Rose application
+final class Item {
+  final String name;
+  int sellIn;
+  int quality;
+
+  Item(this.name, {required this.sellIn, required this.quality});
+
+  @override
+  String toString() => 'Item{name: $name, sellIn: $sellIn, quality: $quality}';
 }