+++ lib/src/namer/approval_number.dart
@@ -10,11 +10,11 @@ class ApprovalNamer {
   String get approved => '$file.approved.txt';
 
   String get approvedFileName =>
-      '${file.split('/').last.split('.').first}.approved.txt';
+      '${file.split('/').last.split('.dart').first}.approved.txt';
 
   // A getter named `received` that returns the string `'file.received.txt'`.
   String get received => '$file.received.txt';
 
   String get receivedFileName =>
-      '${file.split('/').last.split('.').first}.received.txt';
+      '${file.split('/').last.split('.dart').first}.received.txt';
 }